backend/requirements.txt
fastapi
uvicorn
sentence-transformers
keybert
google-generativeai
python-dotenv
numpy
openai
requests